Refactor Django App Plugins to allow for additional features
