+++ lms/djangoapps/grades/apps.py
@@ -7,7 +7,7 @@ Signal handlers are connected here.
 from django.apps import AppConfig
 from django.conf import settings
 from edx_proctoring.runtime import set_runtime_service
-from openedx.core.djangolib.django_plugins import ProjectType, SettingsType, PluginURLs, PluginSettings
+from openedx.core.djangoapps.plugins.constants import ProjectType, SettingsType, PluginURLs, PluginSettings
 
 
 class GradesConfig(AppConfig):
